% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkRegularizedMatrixExistance.R
\name{checkRegularizedMatrixExistance}
\alias{checkRegularizedMatrixExistance}
\title{checkRegularizedMatrixExistance}
\usage{
checkRegularizedMatrixExistance(regOn, mxModelObject, regIndicators)
}
\description{
checks if the matrix that should be regularized actually exists
}
\author{
Jannik Orzek
}
