% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFinalctParameters.R
\name{computeFinalctParameters}
\alias{computeFinalctParameters}
\title{computeFinalctParameters}
\usage{
computeFinalctParameters(
  minimum_criterion,
  ctsemModelObject,
  alpha,
  gamma,
  regOn,
  regIndicators,
  link,
  dt,
  results
)
}
\description{
computes the finale ct parameters
}
\author{
Jannik Orzek
}
