% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeFinalParameters.R
\name{computeFinalParameters}
\alias{computeFinalParameters}
\title{computeFinalParameters}
\usage{
computeFinalParameters(
  minimum_criterion,
  mxModelObject,
  alpha,
  gamma,
  regOn,
  regIndicators,
  results
)
}
\description{
computes the finale parameters
}
\author{
Jannik Orzek
}
