% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRegValueGrid.R
\name{createRegValueGrid}
\alias{createRegValueGrid}
\title{createRegValueGrid}
\usage{
createRegValueGrid(regValues, regOn)
}
\description{
creates a grid of tuning parameters
}
\author{
Jannik Orzek
}
