% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsList.R
\name{createResultsList}
\alias{createResultsList}
\title{createResultsList}
\usage{
createResultsList(RegValuesGrid, autoCV, k)
}
\description{
creates a list to store the results
}
\author{
Jannik Orzek
}
