% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CvRegMultiGroup.R
\name{CvRegMultiGroup}
\alias{CvRegMultiGroup}
\title{CvRegMultiGroup}
\usage{
CvRegMultiGroup(mxModelObjects, regOn, regType = "lasso", regIndicators,
  regValue_start = 0, regValue_end = 1, regValue_stepsize = 0.01,
  autoCV = FALSE, k = 5, manualCV = NULL, zeroThresh = 0.001,
  scaleCV = TRUE)
}
\description{
Note: regmx is based on the R package \pkg{regsem}. Because of the early status of regmx, it is recommended to use regsem instead!
}
\details{
computes cross-validated multi-group model
}
\examples{

}
\author{
Jannik Orzek
}
