% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findMinumumCriterion.R
\name{findMinumumCriterion}
\alias{findMinumumCriterion}
\title{findMinumumCriterion}
\usage{
findMinumumCriterion(results, criterion, convergedSubset, regOn)
}
\description{
returns the best tuning parameter value for a specified criterion
}
\examples{

}
\author{
Jannik Orzek
}
