% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEstimatedParameters.R
\name{getEstimatedParameters}
\alias{getEstimatedParameters}
\title{getEstimatedParameters}
\usage{
getEstimatedParameters(
  regModel,
  alpha,
  gamma,
  regOn,
  regIndicators,
  zeroThresh
)
}
\description{
Note: regmx is based on the R package \pkg{regsem}. Because of the early status of regmx, it is recommended to use regsem instead!
}
\details{
computes number of estimated parameters
}
\examples{

}
\author{
Jannik Orzek
}
