% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFitMeasures.R
\name{getFitMeasures}
\alias{getFitMeasures}
\title{getFitMeasures}
\usage{
getFitMeasures(regModel, alpha, gamma, regOn, regIndicators,
  cvSample = NULL, zeroThresh = 0.001)
}
\description{
Note: regmx is based on the R package \pkg{regsem}. Because of the early status of regmx, it is recommended to use regsem instead!
}
\details{
computes multiple fit measures
}
\examples{

}
\author{
Jannik Orzek
}
