% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getM2LogL_FIML.R
\name{getM2LogL_FIML}
\alias{getM2LogL_FIML}
\title{getM2LogL_FIML}
\usage{
getM2LogL_FIML(ExpCov, ExpMean, ObsData)
}
\arguments{
\item{ExpCov}{Expected covariance matrix}

\item{ExpMean}{Expected means vector}

\item{ObsData}{observed raw data}
}
\description{
Note: laremm is based on the R package \pkg{regsem}. Because of the early status of laremm, it is recommended to use regsem instead!
}
\details{
This function computes the -2log(Likelihood) based on the formula reported in Hirose 2013 for full data
}
\examples{

}
\author{
Jannik Orzek
}
