% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regModel.R
\name{regModel}
\alias{regModel}
\title{regModel}
\usage{
regModel(mxModelObject, regType = "lasso", regOn, regIndicators,
  regValue = 0)
}
\arguments{
\item{mxModelObject}{run mxModel}

\item{regType}{so far only "lasso" and "ridge" implemented}

\item{regOn}{string vector with matrices that should be regularized.}

\item{regIndicators}{list of matrices indicating which parameters to regularize in the matrices provided in regOn. Has to be in the same order as regOn. 1 Indicates a parameter that will be regularized, 0 an unregularized parameter}

\item{regValue}{numeric value of penalty size}
}
\description{
Note: regmx is based on the R package \pkg{regsem}. Because of the early status of regmx, it is recommended to use regsem instead!
regModel creates a regularized model from an mxModel.
}
\examples{
# The following example is taken from the regsem help to demonstrate the equivalence of both methods:

library(lavaan)
library(OpenMx)
# put variables on same scale for regsem
HS <- data.frame(scale(HolzingerSwineford1939[,7:15]))

# define variables:
latent = c("f1")
manifest = c("x1","x2","x3","x4","x5", "x6", "x7", "x8", "x9")

# define paths:
loadings <- mxPath(from = latent, to = manifest, free = c(F,T,T,T,T,T,T,T,T), values = 1)
lcov <- mxPath(from = latent, arrows = 2, free = T, values = 1)
lmanif <- mxPath(from = manifest, arrows =2 , free =T, values = 1)

# define model:
myModel <- mxModel(name = "myModel", latentVars = latent, manifestVars = manifest, type = "RAM",
                   mxData(observed = HS, type = "raw"), loadings, lcov, lmanif,
                   mxPath(from = "one", to = manifest, free = T)
)

fit_myModel <- mxRun(myModel)
round(fit_myModel$A$values,5)

# create regularized model:

selectedA <- matrix(0, ncol = ncol(fit_myModel$A$values), nrow = nrow(fit_myModel$A$values))
selectedA[c(2,3,7,8,9),10] <-1
regIndicators <- list("selectedA" = selectedA)


reg_model <- regModel(mxModelObject = fit_myModel, regType = "lasso", regOn  = "A", regValue = .05
)
fit_reg_model <- mxRun(reg_model)

round(fit_reg_model$BaseModel$A$values,5)

}
\author{
Jannik Orzek
}
