% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CvOptimCtRegModelObject.R
\name{summary.CvOptimCtRegModelObject}
\alias{summary.CvOptimCtRegModelObject}
\title{Summary function for CvOptimRegModelObject}
\usage{
\method{summary}{CvOptimCtRegModelObject}(object, ...)
}
\arguments{
\item{object}{Object from optimRegModel}

\item{...}{additional parameters.}
}
\description{
Summary function for CvOptimRegModelObject
}
