% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CvOptimRegModelObject.R
\name{summary.CvOptimRegModelObject}
\alias{summary.CvOptimRegModelObject}
\title{Summary function for CvOptimRegModelObject}
\usage{
\method{summary}{CvOptimRegModelObject}(object, ...)
}
\arguments{
\item{object}{Object from optimRegModel}

\item{...}{additional parameters.}
}
\description{
Summary function for CvOptimRegModelObject
}
