% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.OptimRegCtModelObject.R
\name{summary.OptimRegCtModelObject}
\alias{summary.OptimRegCtModelObject}
\title{Summary function for OptimRegCtModelObject}
\usage{
\method{summary}{OptimRegCtModelObject}(object, ...)
}
\arguments{
\item{object}{Object from optimRegCtModel}

\item{...}{additional parameters.}
}
\description{
Summary function for OptimRegCtModelObject
}
