% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.OptimRegModelObject.R
\name{summary.OptimRegModelObject}
\alias{summary.OptimRegModelObject}
\title{Summary function for OptimRegModelObject}
\usage{
\method{summary}{OptimRegModelObject}(object, ...)
}
\arguments{
\item{object}{Object from optimRegModel}

\item{...}{additional parameters.}
}
\description{
Summary function for OptimRegModelObject
}
