% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCtFitMeasures.R
\name{getCtFitMeasures}
\alias{getCtFitMeasures}
\title{getCtFitMeasures}
\usage{
getCtFitMeasures(regCtModel, regType, regOn, regIndicators,
  cvSample = NULL, zeroThresh = 0.001)
}
\description{
getCtFitMeasures
}
